Moved Mail tests to a directory
